Transfer ownership to the executor, not back to the owner.
I'm pretty confident this should be transferring ownership to the `_executor`, not the `_owner`.  The `_owner` is the default owner so that would either fail or be a no-op, I'm not sure which.  Also, the fact that executor address is checked against `0` but the owner address is not implies that the transfer should only happen if there is a separate executor.